README.md
angles.py
Given are n different points in 2-dimensional space. Design an O(n^2log n) algorithm that
determines the number of right triangles that can formed by choosing sets of three points.

babysit.py
A sister and brother run a summer babysitting business. They have a lot of job oers,
and are deciding which jobs to take to maximize their income. Each job comes with the
following information: the day of the job, the starting time of the job, the ending time of
the job, the number of children that need to be babysat, and the hourly rate for the job.
Brother and sister can work separate jobs, except for one case: any job involving four or
more children requires them to work together if they take that job. In order to make sure
they have time to sleep, they must end each day by 11pm, and start no earlier than 6am
the next day.
Assuming that sister and brother are presented with n total jobs to choose from, design
an O(n^2) algorithm that determines the most money that they can earn.

fTree.py
Given is a weighted undirected graph G = (V;E) with positive weights and a subset of
its edges F subset E. An F-containing spanning tree of G is a spanning tree that contains all
edges from F (there might be other edges as well). Give an O(mn) algorithm that finds
the cost of the minimum-cost F-containing spanning tree of G.

chessboard.py
Given is a d1*d2 array Chessboard containing only 0's (empty squares) and 1's (occupied positions).
Moreover, there is an unlimited number of dominos, i.e. 1*2 rectangle pieces. Your task is to
decide if it is possible to cover all of the empty squares on the chess board by non-overlapping
dominos (the dominos cannot cover any of the occupied positions and they cannot "stick out" of
the chess board). Your algorithm should run in time O((d1d2)^3).